Modified ratings-data-service to read from database instead of default built in memory
