+++ ratings-data-service/.gitignore
@@ -31,3 +31,6 @@ build/
 
 ### VS Code ###
 .vscode/
+
+# Ignore application.properties
+src/main/resources/application.properties
