--- ratings-data-service/src/main/java/com/example/ratingsservice/models/Rating.java
@@ -1,31 +1,0 @@
-package com.example.ratingsservice.models;
-
-public class Rating {
-
-    private String movieId;
-    private int rating;
-
-    public Rating() {
-    }
-
-    public Rating(String movieId, int rating) {
-        this.movieId = movieId;
-        this.rating = rating;
-    }
-
-    public String getMovieId() {
-        return movieId;
-    }
-
-    public void setMovieId(String movieId) {
-        this.movieId = movieId;
-    }
-
-    public int getRating() {
-        return rating;
-    }
-
-    public void setRating(int rating) {
-        this.rating = rating;
-    }
-}
