+++ ratings-data-service/src/main/java/com/example/ratingsservice/repository/RatingRepository.java
@@ -1,0 +1,9 @@
+package com.example.ratingsservice.repository;
+
+import com.example.ratingsservice.entity.Rating;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+
+public interface RatingRepository extends JpaRepository<Rating, Integer> {
+    List<Rating> findByUserId(String userId);
+}
