--- ratings-data-service/src/main/java/com/example/ratingsservice/resources/RatingsResource.java
@@ -1,24 +1,0 @@
-package com.example.ratingsservice.resources;
-
-import com.example.ratingsservice.models.Rating;
-import com.example.ratingsservice.models.UserRating;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Arrays;
-import java.util.List;
-
-@RestController
-@RequestMapping("/ratings")
-public class RatingsResource {
-
-    @RequestMapping("/{userId}")
-    public UserRating getRatingsOfUser(@PathVariable String userId) {
-        List<Rating> ratings = Arrays.asList(
-                new Rating("550", 4)
-        );
-
-        return new UserRating(ratings);
-    }
-}
