+++ ratings-data-service/src/main/resources/application.properties.example
@@ -2,9 +2,9 @@ spring.application.name=ratings-data-service
 server.port=8083
 
 # MySQL DataSource
-spring.datasource.url=jdbc:mysql://localhost:3306/ratings_db
-spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.url=jdbc:mysql://localhost:3306/your_db_name
+spring.datasource.username=your_username
+spring.datasource.password=your_password
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 # Hibernate
